Move MACAddressField to utilities
