+++ netbox/netbox/filtersets.py
@@ -5,7 +5,6 @@ from django.db import models
 from django_filters.exceptions import FieldLookupError
 from django_filters.utils import get_model_field, resolve_field
 
-from dcim.forms import MACAddressField
 from extras.choices import CustomFieldFilterLogicChoices
 from extras.filters import TagFilter
 from extras.models import CustomField
@@ -13,6 +12,7 @@ from utilities.constants import (
     FILTER_CHAR_BASED_LOOKUP_MAP, FILTER_NEGATION_LOOKUP_MAP, FILTER_TREENODE_NEGATION_LOOKUP_MAP,
     FILTER_NUMERIC_BASED_LOOKUP_MAP
 )
+from utilities.forms import MACAddressField
 from utilities import filters
 
 __all__ = (
